Extract useCandid hook
